Break out stopwatch
